Provide more robust access to the database
Refactors the database connection to be accessible through a singleton and
(more importantly) ensures that, if you change the file backing the database,
you get a new connection.

This was needed because, with the old code, running all the unit tests
(as opposed to just one class at a time) triggered a SQLITE_READONLY_DBMOVED
error. Because the connection was being cached (through the `lazy` keyword),
new tests were reusing the connection from the previous set of tests,
but by that point the file backing that database had been deleted,
resulting in the error.