Update signup with signal-cli's new JSON mode
In the original implementation, we relied on a patched version of signal-cli.
Fortuitously, in the past couple of weeks, signal-cli implemented its own
"native" JSON output format. This commit switches over to using that.

The new version only slightly different from the original, so a couple of
changes had to be made:

1. It prints the JSON to STDOUT instead of a given file
2. The format of the JSON is a bit different